CognitiveServices/requirements.txt
ibm-watson==3.0.4

azure-cognitiveservices-speech==1.5.0
azure-cognitiveservices-language-textanalytics==0.2.0

google-cloud-speech==1.0.0
google-cloud-language==1.2.0

six==1.11.0
Flask==1.0.2
jiwer==1.3.2
pyaudio==0.2.11
websocket-server==0.4


